Refactor protocol to use multiple connections to the MongoDB remote
For standalone MongoDB servers, there is always only one connection.
When working with a cluster ("replica set"), connections are selected
based on the read preference expressed via the connection string.

With a default replicaset size of 3 instances, our initial connect
may result in a connection to a secondary node. This is perfect for
reading if the read preference includes this, but we will need use
the primary for writing, and thus establish a separate connection.

To implement this, the following steps were taken:

* Extract wire connection from protocol implementation. We now have
  a Protocol class (handles multiple connections as well as selecting
  the correct one) as well as a Connection class (a single connection).
  These live in a newly created package com.mongodb.io
* Determine which of the Collection and Database methods perform
  read and which of them perform write operations. Commans are always
  considered to be write operations, aggregation pipelines are checked
  for $out or $merge stages, which are writes, defaulting to read.
* Add a select() method to select (and optionally connect to) one of
  the nodes eligible, handling connection failures. This method will
  be extended in the future to handle connection closed scenarios and
  the implement #9